backend/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
uwsgi==2.0.19.1
djangorestframework==3.12.2
drf_yasg==1.20.0
factory_boy==3.2.0